# src/Wolf.Notification.EmailSender/OpenAPIs/NotifApiSwagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Wolf Notification API",
    "description": "Wolf API for Managing Notification Templates and Messages and for Message-delivery",
    "contact": {
      "name": "API",
      "url": "https://www.zflows.com/"
    },
    "license": { },
    "version": "v2.0"
  },
  "paths": {
...
    "/message/{templateId}/preview": {
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Returns the compiled body and subject of the message.",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "description": "Template ID (GUID)",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Template ID (GUID)",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "nullable": true
              }
            },
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessagePreviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessagePreviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessagePreviewDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/message/{templateName}/preview-by-name": {
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Returns the compiled body and subject of the message for a given Template Name.",
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Template Name",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Template Name"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessagePreviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessagePreviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessagePreviewDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/message/{templateId}/send": {
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Compiles subject and body of message.await Saves to database. Posts event to queue",
        "parameters": [
          {
            "name": "templateId",
            "in": "path",
            "description": "Template ID (GUID)",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Template ID (GUID)",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessageDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessageDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessageDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
    "/message/{templateName}/send-by-name": {
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Compiles subject and body of message.await Saves to database. Posts event to queue",
        "parameters": [
          {
            "name": "templateName",
            "in": "path",
            "description": "Template Name",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Template Name"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Wolf.Notification.Models.SendMessageRequest"
                  }
                ],
                "description": "",
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessageDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessageDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wolf.Notification.Models.MessageDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "OAuth2": [
              "OAuth2"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Wolf.Notification.Models.RecipientDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Wolf.Notification.Models.SendMessageRequest": {
        "type": "object",
        "properties": {
          "tokenValues": {
            "type": "object",
            "additionalProperties": { },
            "description": "Optional dictionary of values to replace in the template",
            "nullable": true
          },
          "providerCode": {
            "type": "string",
            "description": "Optional notification provider code. If not specified the one in template is used. If the field is not specified in the template an error is thrown.",
            "nullable": true
          },
          "sender": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientDto"
              }
            ],
            "description": "Optional Sender name/address. If not specified the one in template is used. If the field is not specified in the template an error is thrown.",
            "nullable": true
          },
          "recipients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWTypeDto"
            },
            "description": "Optional list of Recipients. If not specified the one in template is used. If the field is not specified in the template an error is thrown.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Wolf.Notification.Models.MessagePreviewDto": {
        "type": "object",
        "properties": {
          "provider": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Wolf.Notification.Models.ProviderDto"
              }
            ],
            "nullable": true
          },
          "template": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Wolf.Notification.Models.TemplateIdNameDto"
              }
            ],
            "nullable": true
          },
          "tokenData": {
            "type": "string",
            "nullable": true
          },
          "sender": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientDto"
              }
            ],
            "nullable": true
          },
          "recipients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Wolf.Notification.Models.RecipientWTypeDto"
            },
            "nullable": true
          },
          "subject": {
            "type": "string",
            "nullable": true
          },
          "body": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "description": "Balea Server OpenId Security Scheme",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://devid.mycompany.com/connect/authorize",
            "tokenUrl": "https://devid.mycompany.com/connect/token",
            "scopes": {
              "notif_api": "Wolf Notification Api Audience",
              "openid": "OpenID"
            }
          }
        }
      }
    }
  }
}